requirements.txt
plotly
dash
dash-bootstrap-components
pandas
pandas-gbq
google-cloud-bigquery
requests
python-dotenv
gunicorn